requirements.txt
python-binance
python-telegram-bot
pandas
numpy
ta
python-dotenv
ccxt
requests

